server/requirements.txt
# Door Se Kaam — Python Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
mss>=9.0.0
Pillow>=10.0.0
pyautogui>=0.9.54
pyperclip>=1.8.2
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
cryptography>=41.0.0
